Auto-complete even when hint equals query
_autocomplete checks if the query is not equal to the hint before
setting the input to the first suggestion and firing an event. This
caused an issue for us. The hint is the token, and by default tokens are
all lower cased. If you had typed "foo" and the suggested value was
"Foo", then hit tab, auto-complete would not fire and the input would
remain unchanged with "foo".