Fixed remote dataset returning max of 1 result
We should run the duplication check on the normalized datum rather than
the raw dataset. If we use the raw object that doesn't have a `value`
property (ie. if you're specifying your own `valueKey`),
`that.dupChecker` will always return true and thus we will only ever get
a maximum of one result.