Allow transport url to be a function
Allowing this we would be able to change the URL on the fly in order to
combine more fields to build a query url.

e.g.
I have two autocomplete fields, one is a state, and the other is a
document emitter, but my need is, when the user pick up a state, they
will only see emitters from that state. So the autocomplete call will
look like that:

```
$('#emitter').autocomplete({
   url: function () {
       var state = $("#state").val();

       return '/emitters?q=%QUERY&state=' + state;
   }
});
```
Each time before a request this function will be evaluated, generating
a new url for the request according to what was defined. The only
requirement is this function to return a string.