Changing to use grunt-contrib-concat and intro/outro files for wrapping closure and strict mode
Creating intro.js and outro.js to wrap the dist files (ala jQuery’s
build).  Updating Gruntfile.js to add the intro and outro to the list
of files, add a concat task with targets for bloodhound, typeahead, and
bundle, point uglify src’s at the concat’ed files, and add concat to
the build task.