Set cursor on hover.
Currently, mouse and keyboard interactions are separate. If you navigate using
the keyboard, then hover with the mouse, the cursor remains on the suggestion
the keyboard had navigated to.

With this change, the selected suggestion is set on hover, so hovering and then
pressing a key will use the hovered suggestion as its starting point.