Fixing bug when suggestions length is less than the limit
A bug occurs when the number of suggestions is less than the limit.
Before, async update would give back `suggestions.slice(0, limit -
rendered)` but in cases where number of rendered suggestions is less
than limit, this would result in missing suggestions.

e.g. Say the limit is 5 and the list of suggestions for my query "foo"
comes back as ["food", "fool", "foot"], this would result in appending
`["food", "fool", "foot"].splice(0, 5 - 3) //=> ["food", "fool"]` which
is missing the legitimate suggestion "foot".