Add option to refresh suggestion list on DOM only after async callback is done
On some use cases it can be useful to render the search suggestions after all suggestions are available from synchronous and asynchronous callback of a dataset. So, the flag "updateOnAsync" is introduced to have the possibility to activate the new mode, to render the whole result after asynchronous callback only.

If you want to display the whole result only, then this mode can be used now. Also, if you work with asynchronous data only and will not use a pending template, this mode ensures to prevent closing and reopen the dropdown menu on keypress (which can be misinterpreted as an ugly flickering, while the user types something) provokes by the empty or unused synchronous callback.